refactor round robin LB
